+++ module-auth/src/main/kotlin/com/raillylinker/module_auth/jpa_beans/db1_main/repositories_dsl/Db1_RaillyLinkerCompany_RepositoryDsl.kt
@@ -55,7 +55,7 @@ class Db1_RaillyLinkerCompany_RepositoryDsl(entityManager: EntityManager) {
 
     data class FindAllNowActivateMemberLockInfoOutputVo(
         val uid: Long,
-        val lockReasonCode: Byte,
+        val lockReasonCode: Short,
         val lockReason: String,
         val lockStart: LocalDateTime,
         val lockBefore: LocalDateTime?,
